// src/components/types/FormData.ts
export interface FormData {
    keywords: string;
    keyword_char_name: string;
    keyword_char_status: string;
    keyword_char_spec: string;
    keyword_char_type: string;
    keyword_char_gender: string;
    keyword_loc_name: string;
    keyword_loc_type: string;
    keyword_loc_dimention: string;
    keyword_epi_name: string;
    keyword_epi_epi: string;
};